Add admin upload api
