+++ routers/index.js
@@ -1,5 +1,6 @@
 module.exports = {
   image: require('./image'),
   upload: require('./upload'),
-  proxy: require('./proxy')
+  proxy: require('./proxy'),
+  admin: require('./admin'),
 }
